Removed intro message and added API to set sensor parameters
I removed the introduction message that the Xbee broadcasts when it
connects to the network as I believe this message and the data it
contains are better set from UI connected to the base station instead
of individually programming each sensor. To that end, now every sensor
broadcasts a stream of comma separated sensor values, and it is up to
the base station to make sense of which key each value should be paired
with.

Also, I added an API that reads the first byte of each received message
to set sensor parameters such as synchronization and time period
between readings along with whatever needs may come up in the future.